+++ commit/core.pyx
@@ -21,7 +21,7 @@ import pyximport
 from amico.util import LOG, NOTE, WARNING, ERROR
 
 
-def setup( lmax = 12, ndirs = 32761 ) :
+def setup( lmax=12, ndirs=32761 ) :
     """General setup/initialization of the COMMIT framework.
     
     Parameters
@@ -37,7 +37,7 @@ def setup( lmax = 12, ndirs = 32761 ) :
 
     amico.lut.precompute_rotation_matrices( lmax, ndirs )
 
-def load_dictionary_info(filename):
+def load_dictionary_info( filename ):
     """Function to load dictionary info file
     
     Parameters
@@ -57,6 +57,7 @@ def load_dictionary_info(filename):
             return result_aux
         else:
             return pickle.load( dictionary_info_file )
+
 
 cdef class Evaluation :
     """Class to hold all the information (data and parameters) when performing an
@@ -105,15 +106,15 @@ cdef class Evaluation :
         self.set_config('doNormalizeMaps', False)
 
 
-
     def set_config( self, key, value ) :
         self.CONFIG[ key ] = value
+
 
     def get_config( self, key ) :
         return self.CONFIG.get( key )
 
 
-    def load_data( self, dwi_filename = 'DWI.nii', scheme_filename = 'DWI.scheme', b0_thr = 0 ) :
+    def load_data( self, dwi_filename='DWI.nii', scheme_filename='DWI.scheme', b0_thr=0 ) :
         """Load the diffusion signal and its corresponding acquisition scheme.
 
         Parameters
@@ -215,7 +216,7 @@ cdef class Evaluation :
         self.set_config('ATOMS_path', pjoin( self.get_config('study_path'), 'kernels', self.model.id ))
 
 
-    def generate_kernels( self, regenerate = False, lmax = 12, ndirs = 32761 ) :
+    def generate_kernels( self, regenerate=False, lmax=12, ndirs=32761 ) :
         """Generate the high-resolution response functions for each compartment.
         Dispatch to the proper function, depending on the model.
 
@@ -334,19 +335,18 @@ cdef class Evaluation :
         LOG( '   [ %.1f seconds ]' % ( time.time() - tic ) )
 
 
-    cpdef load_dictionary( self, path, use_mask = False ) :
+    cpdef load_dictionary( self, path, use_all_voxels_in_mask=False ) :
         """Load the sparse structure previously created with "trk2dictionary" script.
 
         Parameters
         ----------
         path : string
             Folder containing the output of the trk2dictionary script (relative to subject path)
-        use_mask : boolean
+        use_all_voxels_in_mask : boolean
             If False (default) the optimization will be conducted only on the voxels actually
-            traversed by tracts. If True, the mask specified in trk2dictionary
-            (i.e. "filename_mask" paramater) will be used instead.
-            NB: if no mask was specified in trk2dictionary, the "tdi" and
-            "mask" masks are equivalent and this parameter is not influent.
+            traversed by tracts. If True, then all voxels present in the mask specified in 
+            trk2dictionary.run(), i.e. "filename_mask" parameter, will be used instead.
+            NB: if no mask was specified in trk2dictionary, this parameter is irrelevant.
         """
         if self.niiDWI is None :
             ERROR( 'Data not loaded; call "load_data()" first' )
@@ -357,7 +357,7 @@ cdef class Evaluation :
         self.set_config('TRACKING_path', pjoin(self.get_config('DATA_path'),path))
 
         # load mask
-        self.set_config('dictionary_mask', 'mask' if use_mask else 'tdi' )
+        self.set_config('dictionary_mask', 'mask' if use_all_voxels_in_mask else 'tdi' )
         mask_filename = pjoin(self.get_config('TRACKING_path'),'dictionary_%s.nii'%self.get_config('dictionary_mask'))
         if not exists( mask_filename ) :
             mask_filename += '.gz'
@@ -387,8 +387,10 @@ cdef class Evaluation :
             ERROR( 'Dictionary is outdated. Execute "trk2dictionary" script first' )
 
         self.DICTIONARY['TRK'] = {}
+        self.DICTIONARY['TRK']['kept']  = np.fromfile( pjoin(self.get_config('TRACKING_path'),'dictionary_TRK_kept.dict'), dtype=np.uint8 )
         self.DICTIONARY['TRK']['norm'] = np.fromfile( pjoin(self.get_config('TRACKING_path'),'dictionary_TRK_norm.dict'), dtype=np.float32 )
         self.DICTIONARY['TRK']['len']  = np.fromfile( pjoin(self.get_config('TRACKING_path'),'dictionary_TRK_len.dict'), dtype=np.float32 )
+        
 
         self.DICTIONARY['IC'] = {}
         self.DICTIONARY['IC']['fiber'] = np.fromfile( pjoin(self.get_config('TRACKING_path'),'dictionary_IC_f.dict'), dtype=np.uint32 )
@@ -481,7 +483,7 @@ cdef class Evaluation :
         LOG( '   [ %.1f seconds ]' % ( time.time() - tic ) )
 
 
-    def set_threads( self, n = None ) :
+    def set_threads( self, n=None ) :
         """Set the number of threads to use for the matrix-vector operations with A and A'.
 
         Parameters
@@ -685,7 +687,7 @@ cdef class Evaluation :
         return self.niiDWI_img[ self.DICTIONARY['MASK_ix'], self.DICTIONARY['MASK_iy'], self.DICTIONARY['MASK_iz'], : ].flatten().astype(np.float64)
 
 
-    def fit( self, tol_fun = 1e-3, tol_x = 1e-6, max_iter = 100, verbose = 1, x0 = None, regularisation = None ) :
+    def fit( self, tol_fun=1e-3, tol_x=1e-6, max_iter=100, verbose=1, x0=None, regularisation=None ) :
         """Fit the model to the data.
 
         Parameters
@@ -740,29 +742,56 @@ cdef class Evaluation :
         LOG( '\n   [ %s ]' % ( time.strftime("%Hh %Mm %Ss", time.gmtime(self.CONFIG['optimization']['fit_time']) ) ) )
 
 
-    def save_results( self, path_suffix = None, save_opt_details = True, save_coeff = False, save_est_dwi = False ) :
+    def get_coeffs( self ):
+        """
+        Returns the coefficients, corresponding to the original optimisation problem,
+        i.e. the input tractogram to trk2dictionary, divided in three classes (ic, ec, iso).
+        """
+        if self.x is None :
+            ERROR( 'Model not fitted to the data; call "fit()" first' )
+
+        nF = self.DICTIONARY['IC']['nF']
+        nE = self.DICTIONARY['EC']['nE']
+        nV = self.DICTIONARY['nV']
+
+        if self.get_config('doNormalizeKernels') :
+            # renormalize the coefficients
+            norm1 = np.repeat(self.KERNELS['wmr_norm'],nF)
+            norm2 = np.repeat(self.KERNELS['wmh_norm'],nE)
+            norm3 = np.repeat(self.KERNELS['iso_norm'],nV)
+            norm_fib = np.kron(np.ones(self.KERNELS['wmr'].shape[0]), self.DICTIONARY['TRK']['norm'])
+            x = self.x / np.hstack( (norm1*norm_fib,norm2,norm3) )
+        else :
+            x = self.x
+
+        offset1 = nF * self.KERNELS['wmr'].shape[0]
+        offset2 = offset1 + nE * self.KERNELS['wmh'].shape[0]
+        kept = np.tile( self.DICTIONARY['TRK']['kept'], self.KERNELS['wmr'].shape[0] )
+        xic = np.zeros( kept.size )
+        xic[kept==1] = x[:offset1]
+        xec = x[offset1:offset2]
+        xiso = x[offset2:]
+
+        return xic, xec, xiso
+
+
+    def save_results( self, path_suffix=None, stat_coeffs='sum', save_est_dwi=False, save_coeff=None, save_opt_details=None ) :
         """Save the output (coefficients, errors, maps etc).
 
         Parameters
         ----------
         path_suffix : string
             Text to be appended to "Results" to create the output path (default : None)
-        save_opt_details : boolean
-            Save everything in a pickle file containing the following list L:
-                L[0]: dictionary with all the configuration details
-                L[1]: np.array obtained through the optimisation process with the normalised kernels
-                L[2]: np.array renormalisation of L[1]
-            (default : True)
-        save_coeff : boolean
-            Save the coefficients related to each compartment in txt files
-            and a pickle file containing the configuration details.
-            (default : False)
+        stat_coeffs : string
+            Stat to be used if more coefficients are estimated for each streamline.
+            Options: 'sum', 'mean', 'median', 'min', 'max', 'all' (default : 'sum')
         save_est_dwi : boolean
             Save the estimated DW-MRI signal (default : False)
+        save_opt_details : boolean
+            DEPRECATED. The details of the optimization and the coefficients are always saved.
+        save_coeff : boolean
+            DEPRECATED. The estimated weights for the streamlines are always saved.
         """
-        if self.x is None :
-            ERROR( 'Model not fitted to the data; call "fit()" first' )
-
         RESULTS_path = 'Results_' + self.model.id
         if path_suffix :
             self.set_config('path_suffix', path_suffix)
@@ -770,6 +799,15 @@ cdef class Evaluation :
 
         LOG( '\n-> Saving results to "%s/*":' % RESULTS_path )
         tic = time.time()
+
+        if self.x is None :
+            ERROR( 'Model not fitted to the data; call "fit()" first' )
+
+        if save_coeff is not None :
+            WARNING('"save_coeff" parameter is deprecated')
+
+        if save_opt_details is not None :
+            WARNING('"save_opt_details" parameter is deprecated')
 
         nF = self.DICTIONARY['IC']['nF']
         nE = self.DICTIONARY['EC']['nE']
@@ -864,13 +902,13 @@ cdef class Evaluation :
         print( '   [ OK ]' )
 
         if self.get_config('doNormalizeMaps') :
-                niiIC = nibabel.Nifti1Image(  niiIC_img  / ( niiIC_img + niiEC_img + niiISO_img + 1e-16), affine )
-                niiEC = nibabel.Nifti1Image(  niiEC_img /  ( niiIC_img + niiEC_img + niiISO_img + 1E-16), affine )
-                niiISO = nibabel.Nifti1Image( niiISO_img / ( niiIC_img + niiEC_img + niiISO_img + 1E-16), affine )
+            niiIC = nibabel.Nifti1Image(  niiIC_img  / ( niiIC_img + niiEC_img + niiISO_img + 1e-16), affine )
+            niiEC = nibabel.Nifti1Image(  niiEC_img /  ( niiIC_img + niiEC_img + niiISO_img + 1E-16), affine )
+            niiISO = nibabel.Nifti1Image( niiISO_img / ( niiIC_img + niiEC_img + niiISO_img + 1E-16), affine )
         else:
-                niiIC = nibabel.Nifti1Image( niiIC_img, affine )
-                niiEC = nibabel.Nifti1Image( niiEC_img, affine )
-                niiISO = nibabel.Nifti1Image( niiISO_img, affine )
+            niiIC = nibabel.Nifti1Image( niiIC_img, affine )
+            niiEC = nibabel.Nifti1Image( niiEC_img, affine )
+            niiISO = nibabel.Nifti1Image( niiISO_img, affine )
 
         nibabel.save( niiIC , pjoin(RESULTS_path,'compartment_IC.nii.gz') )
         nibabel.save( niiEC , pjoin(RESULTS_path,'compartment_EC.nii.gz') )
@@ -879,28 +917,36 @@ cdef class Evaluation :
         # Configuration and results
         print( '\t* Configuration and results:' )
 
-        if save_opt_details:
-            print( '\t\t- results.pickle... ', end='' )
-            sys.stdout.flush()
-            with open( pjoin(RESULTS_path,'results.pickle'), 'wb+' ) as fid :
-                pickle.dump( [self.CONFIG, self.x, x], fid, protocol=2 )
-            print( '[ OK ]' )
+        print( '\t\t- streamline_weights.txt... ', end='' )
+        sys.stdout.flush()
+        xic, _, _ = self.get_coeffs()
+        if stat_coeffs != 'all' and xic.size > 0 :
+            xic = np.reshape( xic, (-1,self.DICTIONARY['TRK']['kept'].size) )
+            if stat_coeffs == 'sum' :
+                xic = np.sum( xic, axis=0 )
+            elif stat_coeffs == 'mean' :
+                xic = np.mean( xic, axis=0 )
+            elif stat_coeffs == 'median' :
+                xic = np.median( xic, axis=0 )
+            elif stat_coeffs == 'min' :
+                xic = np.min( xic, axis=0 )
+            elif stat_coeffs == 'max' :
+                xic = np.max( xic, axis=0 )
+            else :
+                ERROR( 'Stat not allowed. Possible values: sum, mean, median, min, max, all.', prefix='\n' )
+        np.savetxt( pjoin(RESULTS_path,'streamline_weights.txt'), xic, fmt='%.5e' )
+        self.set_config('stat_coeffs', stat_coeffs)
+        print( '[ OK ]' )
 
-        if save_coeff:
-            print( '\t\t- Coefficients txt files... ', end='' )
-            sys.stdout.flush()
-            if len(self.KERNELS['wmr']) > 0 :
-                offset = nF * self.KERNELS['wmr'].shape[0]
-                np.savetxt(pjoin(RESULTS_path,'xic.txt'), x[:offset], fmt='%12.5e')
-            if len(self.KERNELS['wmh']) > 0 :
-                offset = nF * self.KERNELS['wmr'].shape[0]
-                np.savetxt(pjoin(RESULTS_path,'xec.txt'), x[offset:offset+nE*len(self.KERNELS['wmh'])], fmt='%12.5e')
-            if len(self.KERNELS['iso']) > 0 :
-                offset = nF * self.KERNELS['wmr'].shape[0] + nE * self.KERNELS['wmh'].shape[0]
-                np.savetxt(pjoin(RESULTS_path,'xiso.txt'), x[offset:], fmt='%12.5e')
-            with open( pjoin(RESULTS_path,'config.pickle'), 'wb+' ) as fid :
-                pickle.dump( self.CONFIG, fid, protocol=2 )
-            print( '[ OK ]' )
+        # Save to a pickle file the following items:
+        #   item 0: dictionary with all the configuration details
+        #   item 1: np.array obtained through the optimisation process with the normalised kernels
+        #   item 2: np.array renormalisation of coeffs in item 1
+        print( '\t\t- results.pickle... ', end='' )
+        sys.stdout.flush()
+        with open( pjoin(RESULTS_path,'results.pickle'), 'wb+' ) as fid :
+            pickle.dump( [self.CONFIG, self.x, x], fid, protocol=2 )
+        print( '[ OK ]' )
 
         if save_est_dwi :
             print( '\t\t- Estimated signal... ', end='' )